% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_conquest.R
\name{cfb_conquest}
\alias{cfb_conquest}
\title{cfb_conquest}
\usage{
cfb_conquest(season, week, division = "fbs")
}
\arguments{
\item{season}{(Integer required): Year, 4 digit format (YYYY), ~1897 to present.}

\item{week}{(Integer required): Week, values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or eariler).}

\item{division}{(String required): What is the lowest division that should start off owning land? Acceptable inputs include "fbs" or "fcs". \emph{Defaults to "fbs"}.}
}
\value{
Returns data frame prepared for hf_map().
}
\description{
Provides a data frame with current college football conquest data during a particular season and week. The data frame is ready for plotting with hf_map().
}
\examples{
\dontrun{cfb_conquest(season = 2016, week = 8)}
}
