% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cfb_imperialism.R
\name{get_cfb_imperialism}
\alias{get_cfb_imperialism}
\title{get_cfb_imperialism}
\usage{
get_cfb_imperialism(season, week, division = "fbs")
}
\arguments{
\item{season}{(Integer required): Year, 4 digit format (YYYY), ~1897 to present.}

\item{week}{(Integer required): Week, values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or eariler).}

\item{division}{(String required): What is the lowest division that should start off owning land? Acceptable inputs include "fbs" or "fcs". \emph{Defaults to "fbs"}.}
}
\value{
Returns data frame prepared for territorymap().
}
\description{
Provides a data frame with current college football imperialism data during a particular season and week. The data frame is ready for plotting with territorymaps().
}
\examples{
\dontrun{get_cfb_imperialism(season = 2016, week = 8)}
}
