% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cfb_undefeated.R
\name{get_cfb_undefeated}
\alias{get_cfb_undefeated}
\title{get_cfb_undefeated}
\usage{
get_cfb_undefeated(season = 2021, week = 6)
}
\arguments{
\item{season}{(Integer required): Year, 4 digit format (YYYY), ~1897 to present}

\item{week}{(Integer required): Week - values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or eariler)}
}
\value{
Returns data frame prepared for territorymap()
}
\description{
Provides a data frame with information regarding undefeated D1 College Football teams. The data frame is ready for plotting with territorymaps().
}
\examples{
get_cfb_undefeated(season = 2016, week = 8)
}
