% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map_stats.R
\name{get_map_stats}
\alias{get_map_stats}
\title{get_map_stats}
\usage{
get_map_stats(x, continental = TRUE)
}
\arguments{
\item{x}{Data frame created by other territorymap functions or including the following columns:\cr
\cr
\strong{identifier} - identifies each element (ex. school name - Iowa State, Minnesota, Bowling Green).\cr
\cr
\strong{lat} - latitude of element.\cr
\cr
\strong{lng} - longitude of element.\cr

Columns must also be of equal length and match the specified names exactly.}

\item{continental}{(Boolean required): If TRUE (default), only considers the continental United States. If FALSE, also considers Alaska and Hawaii.}
}
\description{
Returns a data frame with summary statistics for each identifier based on their map territory. \cr
Note that this function requires a CENSUS API Key.
Follow the directions \href{https://walker-data.com/tidycensus/reference/census_api_key.html}{here}
to receive a key. \cr
\cr
Output data frame includes four summary statistics:\cr
\cr
\strong{land} - Total land area area (in square meters) within territory controlled by identifier.\cr
\cr
\strong{water} - Total land area area (in square meters) within territory controlled by identifier.\cr
\cr
\strong{domain} - Sum of land and water area (in square meters) within territory controlled by identifier.\cr
\cr
\strong{pop} - Population that resides within territory controlled by identifier.\cr
}
\examples{
\dontrun{cfb_map_stats <- get_map_stats(get_cfb_undefeated(season = 2021, week = 6))}
}
