% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homefield_shiny.R
\name{homefield_shiny}
\alias{homefield_shiny}
\title{homefield_shiny}
\usage{
homefield_shiny()
}
\description{
Opens the homefield Shiny App and features college football homefield maps.
}
\examples{
\dontrun{homefield::homefield_shiny()}
}
