% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/territorymap.R
\name{territorymap}
\alias{territorymap}
\title{territorymap}
\usage{
territorymap(x, threshold = 10000, output_file)
}
\arguments{
\item{x}{data frame with specific columns TBD}

\item{threshold}{area in km^2 for secondary logos}

\item{output_file}{file path for output .png}
}
\description{
territorymap
}
\examples{
cfb_data <- get_cfb_undefeated(season = 2016, week = 6)
territorymap(x = cfb_data, output_file = "C:/Users/darthvader/Downloads")
}
